Use int64_t because everyone else does. And it's more useful.

+++ include/platform/mir/graphics/frame.h
@@ -24,7 +24,7 @@
 
 namespace mir { namespace graphics {
 
-inline uint64_t current_ust(clockid_t clock_id)
+inline int64_t current_ust(clockid_t clock_id)
 {
     struct timespec now;
     clock_gettime(clock_id, &now);
@@ -40,28 +40,28 @@ inline uint64_t current_ust(clockid_t clock_id)
  *   EGL: https://bugs.chromium.org/p/chromium/issues/attachmentText?aid=178027
  *   WGL: https://www.opengl.org/registry/specs/OML/wgl_sync_control.txt
  *
- * The simplistic uint64 types are intentional as all these values originate
+ * The simplistic int64 types are intentional as all these values originate
  * from the display hardware and must be passed unmodified from the server to
  * clients, and clients of clients (even clients of clients of clients like a
  * GLX app under Xmir under Unity8 under USC)!
  */
 struct Frame
 {
-    uint64_t msc = 0;  /**< Media Stream Counter: Counter of the frame
+    int64_t msc = 0;   /**< Media Stream Counter: Counter of the frame
                             displayed by the output (or as close to it as
                             can be estimated). */
     clockid_t clock_id = CLOCK_MONOTONIC;
                        /**< The system clock identifier that 'ust' is measured
                             by. Display drivers usually use CLOCK_MONOTONIC
                             but not always. */
-    uint64_t ust = 0;  /**< Unadjusted System Time in microseconds of the frame
+    int64_t ust = 0;   /**< Unadjusted System Time in microseconds of the frame
                             displayed by the output, relative to:
                               clock_gettime(frame.clock_id, &tp);
                             or
                               mir::graphics::current_ust(frame.clock_id)
                             It is crucial that all processes use the same
                             clock_id when comparing Frame information. */
-    uint64_t min_ust_interval = 0;
+    int64_t min_ust_interval = 0;
                        /**< The shortest time possible to the next frame you
                             should expect. This value may change over time and
                             should not be assumed to remain constant,
@@ -70,7 +70,7 @@ struct Frame
 
     int64_t age() const // might be negative in some cases (in the future)
     {
-        return static_cast<int64_t>(current_ust(clock_id) - ust);
+        return current_ust(clock_id) - ust;
     }
 
     Frame predict_next() const
